+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### New
+
+- Support for `video` tag
+
+### Changed
+
+- Improved responsiveness when used with `iframe` tag
+
 ## [3.0.4] - 2017-08-08
 
 ### Changed